backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
yfinance==0.2.32
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
websocket-client==1.7.0
pyjwt==2.8.0
werkzeug==3.0.1
